requirements.txt
PyYAML==6.0.1
g4f==0.1.8.6
pandas==2.1.2
requests==2.31.0
